fix(grpc): R1-R3 review fixes — gRPC IPC migration hardening
review-fix-loop 3ラウンド (R1: 35件 / R2: 11件 / R3: 4件) の修正を反映。

主な変更:
- _GrpcAttachClient.wait_for: Error/EngineBusy/timeout を ConnectionError/BusyError/TimeoutError に変換、
  request_id フィルタで別クライアント宛エラーを無視、pending バッファで無関係イベントを保存し正常終了時に requeue
- _GrpcAttachClient.events(): Error イベントに request_id フィルタを追加（wait_for との対称性）
- _async_main: post-handshake ログに例外メッセージを追記
- server_grpc.py: schemas.py から SCHEMA 定数を import しないローカル定数化、
  schema_minor 不一致 warning、MAX_CONNECTIONS 超過 warning、ignore_unknown_fields=False
- _Broadcaster: ServerConnection 依存を Any 型に緩和し _GrpcSessionKey に対応
- grpc_transport.rs: ReadyResponse schema_major クロスチェック、NotifyOnDrop RAII ガード、
  Unspecified バリアント warn + drop (SignalKind は None 返却、PositionType は return None)
- SCHEMA_MAJOR/MINOR: u16 → u32 (proto uint32 と一致)
- check_schema_parity.py: repeated フィールド対称確認、oneof 存在確認、
  ブレース深さカウンタ (extract_message_block)、フィールド番号 gap 検出
- proto-lint.yml: buf コマンドをルート buf.yaml 対応に修正
- テスト追加: handshake/start_or_attach シナリオ、R2-H1/M1/M2/M3/H2/H3 各テスト

cargo test --workspace / uv run pytest python/tests/ → 全 PASS

+++ engine-client/src/connection.rs
@@ -65,7 +65,7 @@ impl EngineConnection {
         target: &str,
         token: &str,
         mode: crate::dto::AppMode,
-        schema_major_override: u16,
+        schema_major_override: u32,
     ) -> Result<Self, EngineClientError> {
         let (sender, events, closed, capabilities) =
             crate::grpc_transport::start_grpc_session_with_schema(
