+++ engine-client/src/dto.rs
@@ -1,7 +1,7 @@
 /// IPC Data Transfer Objects for the Rust ↔ Python data engine protocol.
 ///
 /// Commands flow Rust → Python; Events flow Python → Rust.
-/// Both are transported as JSON text frames over a local WebSocket.
+/// Both are transported as protobuf frames over a local gRPC stream.
 use serde::{Deserialize, Serialize};
 
 /// N1.13 / R1b H-E: 起動時固定モード。CLI `--mode {live|replay}` で指定する。
@@ -50,8 +50,8 @@ impl AppMode {
 #[serde(tag = "op")]
 pub enum Command {
     Hello {
-        schema_major: u16,
-        schema_minor: u16,
+        schema_major: u32,
+        schema_minor: u32,
         client_version: String,
         token: String,
         /// N1.13 / R1b H-E: 起動時に固定する mode (`AppMode::Live` | `AppMode::Replay`).
@@ -980,8 +980,8 @@ impl std::fmt::Display for CurrentEngineState {
 #[serde(tag = "event")]
 pub enum EngineEvent {
     Ready {
-        schema_major: u16,
-        schema_minor: u16,
+        schema_major: u32,
+        schema_minor: u32,
         engine_version: String,
         engine_session_id: String,
         // `#[serde(default)]` is a defensive read for older engines that may
